+++ src/extension-test/unit/util/get-text-test.ts
@@ -9,7 +9,7 @@ describe('getTopLevelFunction', () => {
         const selDoc: mock.MockDocument = docFromTextNotation('(foo bar)•(deftest |a-test|•  (baz gaz))');
         expect(getText.currentTopLevelFunction(doc)).toEqual([[selDoc.selectionLeft, selDoc.selectionRight], 'a-test']);
     });
-    
+
     it('Finds top level function when nested', () => {
         const doc: mock.MockDocument = docFromTextNotation('(foo bar)•(with-test•  (deftest a-test•    (baz |gaz)))');
         const selDoc: mock.MockDocument = docFromTextNotation('(foo bar)•(with-test•  (deftest |a-test|•    (baz gaz)))');
@@ -23,4 +23,20 @@ describe('getTopLevelFunction', () => {
         expect(getText.currentTopLevelFunction(doc)).toEqual([[selDoc.selectionLeft, selDoc.selectionRight], 'a-test']);
     });
 
+});
+
+describe('getTopLevelForm', () => {
+    it('Finds top level form', () => {
+        const doc: mock.MockDocument = docFromTextNotation('(foo bar)•(deftest a-test•  (baz |gaz))');
+        const selDoc: mock.MockDocument = docFromTextNotation('(foo bar)•|(deftest a-test•  (baz gaz))|');
+        expect(getText.currentTopLevelForm(doc)).toEqual([[selDoc.selectionLeft, selDoc.selectionRight], '(deftest a-test\n  (baz gaz))']);
+    });
+});
+
+describe('getTopLevelFormToCursor', () => {
+    it('Finds top level form from start to cursor', () => {
+        const doc: mock.MockDocument = docFromTextNotation('(foo bar)•(deftest a-test•  [baz |gaz])');
+        const selDoc: mock.MockDocument = docFromTextNotation('(foo bar)•|(deftest a-test•  [baz |gaz])');
+        expect(getText.currentTopLevelFormToCursor(doc)).toEqual([[selDoc.selectionLeft, doc.selectionRight], '(deftest a-test\n  [baz ])']);
+    });
 });