+++ src/util/cursor-get-text.ts
@@ -1,5 +1,5 @@
 /* Functions for getting text given only cursor-doc related input
- * Can be unit tested since vscode and stuff is not imported 
+ * Can be unit tested since vscode and stuff is not imported
 */
 
 import { LispTokenCursor } from "../cursor-doc/token-cursor";
@@ -22,6 +22,28 @@ export function currentTopLevelFunction(doc: EditableDocument): RangeAndText {
                 break;
             }
         }
+    }
+    return [undefined, ''];
+}
+
+export function currentTopLevelForm(doc: EditableDocument): RangeAndText {
+    const defunCursor = doc.getTokenCursor(0);
+    const defunRange = defunCursor.rangeForDefun(doc.selection.active);
+    return defunRange ? [defunRange, doc.model.getText(...defunRange)] : [undefined, ''];
+}
+
+export function currentTopLevelFormToCursor(doc: EditableDocument): RangeAndText {
+    const defunCursor = doc.getTokenCursor(0);
+    const defunRange = defunCursor.rangeForDefun(doc.selection.active);
+    if (defunRange) {
+        const textUpToCursor = doc.model.getText(defunRange[0], doc.selection.active);
+        const closeBrackets: string[] = [];
+        const bracketCursor = doc.getTokenCursor(doc.selection.active);
+        while (bracketCursor.offsetStart !== defunRange[1] && bracketCursor.forwardList() && bracketCursor.upList()) {
+            closeBrackets.push(bracketCursor.getPrevToken().raw);
+        }
+        const range: [number, number] = [defunRange[0], doc.selection.active];
+        return [range, doc.model.getText(...range) + closeBrackets.join('')];
     }
     return [undefined, ''];
 }